+++ src/main/java/com/popeffects/client/PopEffectsClient.java
@@ -5,7 +5,7 @@ import com.popeffects.config.ConfigManager;
 import com.popeffects.effect.EffectManager;
 import com.popeffects.effect.EffectRenderer;
 import com.popeffects.effect.PopCounter;
-import com.popeffects.effect.PopRenderLayers;
+import com.popeffects.compat.EffectLayers;
 import com.popeffects.trigger.PopTriggers;
 
 import net.fabricmc.api.ClientModInitializer;
@@ -21,7 +21,7 @@ public final class PopEffectsClient implements ClientModInitializer {
 
 		// Muss vor dem ersten Uebersetzen der Shader passieren, deshalb gleich
 		// hier und nicht erst beim ersten Effekt.
-		PopRenderLayers.init();
+		EffectLayers.init();
 
 		EffectRenderer.register();
 		PopEffectsKeybinds.register();